roles management

+++ src/Api/Controllers/AdminController.cs
@@ -9,18 +9,24 @@ namespace Api.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
-public class AdminController(ICourseService courseService,
-    IUserService userService) : ControllerBase
+public class AdminController(
+    ICourseService courseService,
+    IUserService userService,
+    ITokenService tokenService) : ControllerAuth(tokenService)
 {
     [HttpGet("users")]
     public async Task<IActionResult> GetUsers()
     {
+        if (GetRoleFromBearerToken()!= Role.Admin) return Unauthorized();
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         return base.Ok(await userService.GetAllAsync());
     }
 
     [HttpDelete("users/{id:int}")]
     public async Task<IActionResult> DeleteUser(int id)
     {
+        if (GetRoleFromBearerToken()!= Role.Admin) return Unauthorized();
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         bool b = await userService.DeleteByIdAsync(id);
         return b? base.Ok() : base.BadRequest();
     }
@@ -28,6 +34,8 @@ public class AdminController(ICourseService courseService,
     [HttpPut("promote")]
     public async Task<IActionResult> Promote([FromBody] UserPromoteDto userPromote)
     {
+        if (GetRoleFromBearerToken()!= Role.Admin) return Unauthorized();
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         bool b = await userService.PromoteAsync(userPromote);
         return b ? base.Ok() : base.BadRequest();
     }
@@ -35,6 +43,8 @@ public class AdminController(ICourseService courseService,
     [HttpPost("enroll")]
     public async Task<IActionResult> EnrollStudentToCourse([FromBody] EnrollDto dto)
     {
+        if (GetRoleFromBearerToken()!= Role.Admin) return Unauthorized();
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         Role? role = await userService.GetRoleAsync(dto.UserId);
         if (role != Role.Student) return BadRequest();
         return await courseService.EnrollStudentToCourseAsync(dto.CourseId, dto.UserId) ? Ok() : BadRequest();
@@ -43,6 +53,8 @@ public class AdminController(ICourseService courseService,
     [HttpDelete("enroll")]
     public async Task<IActionResult> RemoveStudentFromCourse([FromBody] EnrollDto dto)
     {
+        if (GetRoleFromBearerToken()!= Role.Admin) return Unauthorized();
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         Role? role = await userService.GetRoleAsync(dto.UserId);
         if (role != Role.Student) return BadRequest();
         return await courseService.RemoveStudentToCourseAsync(dto.CourseId, dto.UserId)? Ok() : BadRequest();
@@ -51,6 +63,8 @@ public class AdminController(ICourseService courseService,
     [HttpPost("assign")]
     public async Task<IActionResult> AssignInstructor([FromBody] EnrollDto dto)
     {
+        if (GetRoleFromBearerToken()!= Role.Admin) return Unauthorized();
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         Role? role = await userService.GetRoleAsync(dto.UserId);
         if (role != Role.Instructor) return BadRequest();
         return await courseService.AssignInstructorAsync(dto.CourseId, dto.UserId) ? Ok() : BadRequest();
@@ -59,6 +73,8 @@ public class AdminController(ICourseService courseService,
     [HttpDelete("assign")]
     public async Task<IActionResult> RemoveInstructor([FromBody] EnrollDto dto)
     {
+        if (GetRoleFromBearerToken()!= Role.Admin) return Unauthorized();
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         Role? role = await userService.GetRoleAsync(dto.UserId);
         if (role != Role.Instructor) return BadRequest();
         // instructors can only remove instructor from their own courses
